requirements.txt
streamlit
pymongo
playwright
python-dotenv
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
